feat: Add useResponsivePageSize.js as a custom hook
